% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAE_t.R
\name{BAET.sim}
\alias{BAET.sim}
\title{BAET simulator: this function simulates a Bayesian adaptive enrichment trial}
\usage{
BAET.sim(
  nt,
  ns,
  ss.interim.es,
  response.type = "gaussian",
  sig.e = 10,
  mean.response = matrix(c(seq(nt * ns)), nrow = nt, ncol = ns, byrow = F),
  prob.subpopulation = rep(1/ns, ns),
  prob.trtarm = rep(1/nt, nt),
  maxN = 300,
  upper = rep(0.9, ns),
  lower = rep(0.1, ns),
  rar = F,
  rarmin.p = 0.1,
  rarmax.p = 0.9,
  MOR = rep(-Inf, ns),
  prob.MOR = rep(0.1, ns),
  N.MCMC = 3000,
  prior.cov = diag(100, ns * nt),
  prior.mean = rep(0, ns * nt)
)
}
\arguments{
\item{nt}{number of treatment arms}

\item{ns}{number of subgroups}

\item{ss.interim.es}{A list containing the g vectors, with gth vector indicating the accumulated number of sample sizes for each interim analysis for gth subpopulation}

\item{response.type}{either 'binary'(probability), 'count'(lambda) or 'gaussian'}

\item{mean.response}{vector of mean responses: a nt (row) * ns (col) matrix, with row number indicates treatment and col number indicates subgroup}

\item{prob.subpopulation}{the probability of a subject coming from one specific subpopulation. default: rep (1/ns, ns)}

\item{prob.trtarm}{the (initial) probability of a subject being assigned to a specific treatment arm. default: rep (1/nt, nt)}

\item{maxN}{the maximum sample size, trial will stop when achieving this number}

\item{upper}{upper probability threshold to claim superiority (in a specific subgroup); the treatment arm wins if it goes above this threshold}

\item{lower}{lower probability threshold to claim superiority (in a specific subgroup); the treatment arm will be dropped if it goes below this threshold}

\item{rar}{whether using responsive adaptive randomization (rar)}

\item{rarmin.p}{the minimum randomization probability under rar  default:0.1}

\item{rarmax.p}{the maximum randomization probability under rar  default:0.9}

\item{MOR}{the minimum meaningful outcome threshold for each subgroup}

\item{prob.MOR}{the probability threshold of being larger than the MOR, treatment arms below this threshold will be dropped}

\item{N.MCMC}{Number of MCMC samples}

\item{prior.cov}{the prior covariance for a multivariate normal prior}

\item{prior.mean}{the prior mean for a multivariate normal prior}
}
\value{
n.analysis: number of analysis conducted to end the whole trial

trt_sub: simulated treatment arm allocation (1st column) and subgroup (2nd column)

ss.sub.trt: a ‘ns’ * ‘nt’ matrix storing the sample size for each subpopulation in each treatment arm, with the row number indicating the subpopulation and the column number indicating the treatment arm.

est: treatment effect estimation(posterior mean) and the 95% Credible interval bounds for each treatment arm in each subgroup

powerind: power indicator of whether the best treatment arms is correctly selected in each subgroup

y: simlulated outcome

n_terminate: the total sample size consumed when trial ends

prob_sup_minioutcome: the probability of a treatment large than the MOR

prob_superiority: the probability of a treatment being the best among each subgroup
}
\description{
BAET simulator: this function simulates a Bayesian adaptive enrichment trial
}
