% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_BayesAET-package.R
\docType{package}
\name{BayesAET-package}
\alias{BayesAET}
\alias{BayesAET-package}
\title{BayesAET: Bayesian Adaptive Enrichment Trial Simulation Tools}
\description{
This package provides tools to simulate and design Bayesian adaptive enrichment trials with pre-specified subgroups.
}
