% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAE_t.R
\name{Multi.BAET}
\alias{Multi.BAET}
\title{Multi.BAET: this function simulates the properties of Bayesian adaptive enrichment trial}
\usage{
Multi.BAET(
  n.sim,
  n.cores = detectCores()/2,
  nt = 3,
  ns = 2,
  ss.interim.es,
  response.type = "gaussian",
  sig.e = 10,
  mean.response = matrix(c(8, 12, 14, 5, 8, 11), nrow = nt, ncol = ns, byrow = F),
  prob.subpopulation = c(0.6, 0.4),
  prob.trtarm = rep(1/nt, nt),
  maxN = 300,
  upper = rep(0.9, ns),
  lower = rep(0.1, ns),
  rar = F,
  rarmin.p = 0.1,
  rarmax.p = 0.9,
  MOR = rep(-Inf, ns),
  prob.MOR = rep(0.1, ns),
  N.MCMC = 5000,
  prior.cov = diag(25, ns * nt),
  prior.mean = c(8, 12, 14, 5, 8, 11)
)
}
\arguments{
\item{n.sim}{number of simulations}

\item{n.cores}{number of cores for computation, default is half of the total cores}

\item{nt}{number of treatment arms}

\item{ns}{number of subgroups}

\item{ss.interim.es}{A list containing the g vectors, with gth vector indicating the accumulated number of sample sizes for each interim analysis for gth subpopulation}

\item{response.type}{either 'binary'(probability), 'count'(lambda) or 'gaussian'}

\item{mean.response}{vector of mean responses: a nt (row) * ns (col) matrix, with row number indicates treatment and col number indicates subgroup}

\item{prob.subpopulation}{the probability of a subject coming from one specific subpopulation. default: rep (1/ns, ns)}

\item{prob.trtarm}{the (initial) probability of a subject being assigned to a specific treatment arm. default: rep (1/nt, nt)}

\item{maxN}{the maximum sample size, trial will stop when achieving this number}

\item{upper}{upper probability threshold to claim superiority (in a specific subgroup); the treatment arm wins if it goes above this threshold}

\item{lower}{lower probability threshold to claim superiority (in a specific subgroup); the treatment arm will be dropped if it goes below this threshold}

\item{rar}{whether using responsive adaptive randomization (rar)}

\item{rarmin.p}{the minimum randomization probability under rar  default:0.1}

\item{rarmax.p}{the maximum randomization probability under rar  default:0.9}

\item{MOR}{the minimum meaningful outcome threshold for each subgroup}

\item{prob.MOR}{the probability threshold of being larger than the MOR, treatment arms below this threshold will be dropped}

\item{N.MCMC}{Number of MCMC samples}

\item{prior.cov}{the prior covariance for a multivariate normal prior}

\item{prior.mean}{the prior mean for a multivariate normal prior}
}
\value{
est: the posterior mean of each treatment (col) in each subpopulation (row); e.g. est[1,2] is treatment 2 in subpopulation 1

sd: the posterior sd of each treatment (col) in each subpopulation (row)

ss.sub.dist: the sample size distribution for each subpopulation (row)

ss.sub.trt.mean: a ‘ns’ * ‘nt’ matrix storing the averaged sample size for each subpopulation in each treatment arm, with the row number indicating the subpopulation and the column number indicating the treatment arm.

ss.sub.mean: the expected (mean) sample size for each subpopulation

ss.t.dist: the sample size distribution for the whole trial

ss.t.mean: the mean total sample size for the whole trial

power.sub: the power for each subpopulation

computation.time: the overall simulation time used
}
\description{
Multi.BAET: this function simulates the properties of Bayesian adaptive enrichment trial
}
