% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAE_t.R
\name{Multi.BAET}
\alias{Multi.BAET}
\title{Multi.BAET: this function simulates the properties of Bayesian adaptive enrichment trial}
\usage{
Multi.BAET(
  n.sim,
  n.cores = detectCores()/2,
  nt = 3,
  ns = 2,
  ss.interim.es,
  response.type = "gaussian",
  sig.e = 10,
  mean.response = matrix(c(8, 12, 14, 5, 8, 11), nrow = nt, ncol = ns, byrow = F),
  prob.subpopulation = c(0.6, 0.4),
  prob.trtarm = rep(1/nt, nt),
  maxN = 300,
  upper = rep(0.9, ns),
  lower = rep(0.1, ns),
  rar = F,
  rarmin.p = 0.1,
  rarmax.p = 0.9,
  MOR = rep(-Inf, ns),
  prob.MOR = rep(0.1, ns),
  N.MCMC = 5000,
  prior.cov = diag(25, ns * nt),
  prior.mean = c(8, 12, 14, 5, 8, 11)
)
}
\arguments{
\item{n.sim}{The number of simulations}

\item{n.cores}{The number of cores used for computation, default is half of the total cores}

\item{nt}{The number of treatment arms.}

\item{ns}{The number of subpopulations.}

\item{ss.interim.es}{A list containing g vectors, with g-th vector indicating the accumulated sample sizes at each interim analysis for g-th subpopulation.}

\item{response.type}{The outcome type, one of “gaussian”, “binary” or “count”.}

\item{sig.e}{The standard deviation of the error term when generating a gaussian outcome. The default value is 1. This parameter is only relevant when the “response.type” is set as “gaussian”.}

\item{mean.response}{A ‘nt’ * ‘ns’  matrix of the assumed true mean responses for each treatment (row number) in each subpopulation (column number). For binary and count outcomes, the inputs are the probabilities and rates, respectively.}

\item{prob.subpopulation}{A vector of length ‘ns’ indicating the probabilities of a participant belonging to each of the subpopulations. The default is an equal probability for each subpopulation.}

\item{prob.trtarm}{A vector of length ‘nt’ indicating the (initial) probability of a participant being assigned to each treatment arm. The default is an equal probability to each treatment arm.}

\item{maxN}{The maximum sample size.}

\item{upper}{A vector of length ‘ns’ indicating the posterior probability threshold above which a treatment will be declared best (superiority) for each subpopulation. The default value is 0.90.}

\item{lower}{A vector of length ‘ns’ indicating the posterior probability threshold below which a treatment will be dropped for each subpopulation. The default value is 0.10.}

\item{rar}{A logical indicator of whether using responsive adaptive randomization (RAR). The default setting is ‘false’, i.e., balanced randomization.}

\item{rarmin.p}{The minimum randomization probability under ‘rar’. The default value is 0.1.}

\item{rarmax.p}{The maximum randomization probability under ‘rar’. The default value is 0.9.}

\item{MOR}{A vector of length ‘ns’ indicating the minimum outcome requirement threshold for each subpopulation. The default value is ‘-Inf’.}

\item{prob.MOR}{A vector of length ‘ns’ indicating the posterior probability threshold below which a treatment will be dropped due to the outcome not achieving the ‘MOR’ for each subpopulation. The default value is 0.10.}

\item{N.MCMC}{The number of MCMC samples (excluding burn in samples)}

\item{prior.cov}{A (‘nt’ * ‘ns’) by (‘nt’ * ‘ns’) square matrix for the prior covariance of a multivariate normal prior for the mean responses. The default is a diagonal matrix with ‘100’ on the diagonal.}

\item{prior.mean}{A vector of length ‘nt’ * ‘ns’ for prior mean of a multivariate normal prior for the mean responses. The default is a vector of zeros.}

\item{shape}{The shape parameter for the gamma distribution for the standard deviation of the error term, i.e., “sig.e”. The default value is 0.001. This parameter is only relevant when the “response.type” is set as “gaussian”.}

\item{rate}{The rate parameter for the gamma distribution for the standard deviation of the error term, i.e., “sig.e”. The default value is 0.001. This parameter is only relevant when the “response.type” is set as “gaussian”.}
}
\value{
est.mean: A ‘ns’ * ‘nt’ matrix storing the mean of the posterior mean, with row number indicating subpopulation and column number indicating treatment arm

est.sd: A ‘ns’ * ‘nt’ matrix storing the mean of the posterior SD, with row number indicating subpopulation and column number indicating treatment arm

ss.sub.trt.mean: A ‘ns’ * ‘nt’ matrix storing the averaged sample size for each subpopulation in each treatment arm, with the row number indicating the subpopulation and the column number indicating the treatment arm.

ss.sub.dist: A ‘ns’ * ‘n.sim’ matrix storing the sample size consumed for each subpopulation in each run, with row number indicating subpopulation and column number indicating the specific run

ss.sub.mean: A vector length ‘ns’ indicating the average sample size consumed in each subpopulation

ss.t.dist: A ‘n.sim’ vector storing the overall sample size consumed in each run

ss.t.mean: A number of average overall sample size consumed

power.sub: A vector length ‘ns’ indicating the (empirical) power estimated in each subpopulation

computation.time: Time consumed for computation
}
\description{
Multi.BAET: this function simulates the properties of Bayesian adaptive enrichment trial
}
\examples{
nt=3
ns=2
Multi.BAET(n.sim = 5,
           n.cores=3,
           nt=3, ns=2,
           ss.interim.es = list(c(30, 50, 100), c(20, 60, 100)),
           response.type = "gaussian",
           sig.e = 10,
           mean.response = matrix(c(seq(nt*ns)),nrow = nt, ncol = ns, byrow =F),
           prob.subpopulation = rep (1/ns, ns),
           prob.trtarm = rep (1/nt, nt),
           maxN = 100,
           upper = rep (1, ns),
           lower = rep (0, ns),
           rar = F,
           rarmin.p = 0.1,
           rarmax.p = 0.9,
           MOR = rep(0, ns),
           prob.MOR = rep(0.1, ns),
           N.MCMC = 3000,
           prior.cov = diag(25, ns*nt),
           prior.mean = rep(0, ns*nt))

}
